feat: mocks for database

+++ src/main/kotlin/com/example/modules/users/data/repository/UserRepositoryImpl.kt
@@ -15,7 +15,7 @@ class UserRepositoryImpl(): UserRepository {
             name = userModel.name.value
             email = userModel.email.value
             password = userModel.password.value
-            houseId = userModel.houseId
+            houseId = userModel.houseId.toInt()
         }
     }
 
@@ -48,5 +48,5 @@ private fun UserEntity.toModel() = UserModel(
     email = Email(email),
     name = Name(name),
     password = Password(password),
-    houseId = houseId
+    houseId = houseId.toLong()
 )
