+++ src/main/kotlin/com/example/modules/users/domain/model/UserModel.kt
@@ -8,7 +8,7 @@ data class UserModel(
     val name: Name,
     val email: Email,
     val password: Password,
-    val houseId: Int
+    val houseId: Long
 ){
     @JvmInline @Serializable
     value class Name(val value: String){
